Adding another status result.
To maintain a one to one correspondence to the cublas API, this flag "HIPBLAS_STATUS_ARCH_MISMATCH" must also be added. While code that may report hipblasStatus_t as an architectural mismatch may not exist, adding many small incremental additions such as this helps to make transpilation simpler for the end user.